requirements.txt
SQLAlchemy
python-dotenv
pymysql
cryptography